lib/ffmpeg: add common transcoder base
Add a transcoder base class for ffmpeg to share
common code between qsv and vaapi elements.
